src/requirements.txt
Flask==3.1.2
gpiozero==2.0.1
Jinja2==3.1.6
Pillow==12.0.0
python_crontab==3.3.0
Requests==2.32.5
selenium==4.37.0
spidev==3.8
waitress==3.0.2
Werkzeug==3.1.3
